(refactor) Moves `MockAPIOrderBookDataSource` to the `test` directory.
